refactor: move auth components to the auth package chore: update imports for auth components
