+++ packages/auth/src/components/forms/phone-identifier.tsx
@@ -1,16 +1,16 @@
 "use client";
 
-import { type } from "arktype";
-import * as React from "react";
-import { FieldError } from "../../core/field";
+import type { AuthResult } from "@rectangular-labs/auth/adapter/types";
 import {
   clearFormError,
   handleFormResultError,
   toFieldErrors,
   useAppForm,
-} from "../../ui/tanstack-form";
+} from "@rectangular-labs/ui/components/tanstack-form";
+import { FieldError } from "@rectangular-labs/ui/core/field";
+import { type } from "arktype";
+import * as React from "react";
 import { PhoneFieldGroup } from "../field-groups/phone";
-import type { AuthResult } from "@rectangular-labs/auth/adapter/types";
 
 // ─── Props ───────────────────────────────────────────────────────────────────
