+++ packages/ui/src/components/icon.tsx
@@ -1,13 +1,19 @@
 import {
   ArrowDownIcon,
   ArrowUpIcon,
+  BuildingsIcon,
+  CaretUpDownIcon,
+  CheckIcon,
   XIcon as CloseIcon,
   DotIcon,
   EyeIcon,
   EyeSlashIcon,
   FileIcon,
+  FingerprintSimpleIcon,
   ListIcon,
   MoonIcon,
+  PaperPlaneTiltIcon,
+  SpinnerIcon,
   SunIcon,
   TerminalIcon,
   WarningCircleIcon,
@@ -18,12 +24,18 @@ type IconProps = React.HTMLAttributes<SVGElement>;
 
 export const ArrowDown = ArrowDownIcon;
 export const ArrowUp = ArrowUpIcon;
+export const Buildings = BuildingsIcon;
+export const CaretUpDown = CaretUpDownIcon;
+export const Check = CheckIcon;
 export const Dot = DotIcon;
 export const EyeOff = EyeSlashIcon;
 export const EyeOn = EyeIcon;
 export const File = FileIcon;
+export const Fingerprint = FingerprintSimpleIcon;
 export const Menu = ListIcon;
 export const Moon = MoonIcon;
+export const PaperPlaneTilt = PaperPlaneTiltIcon;
+export const Spinner = SpinnerIcon;
 export const Sun = SunIcon;
 export const Terminal = TerminalIcon;
 export const WarningCircle = WarningCircleIcon;
